requirements.txt
transformers
matrix-nio[e2e]
torch
nltk
